+++ CHANGELOG.md
@@ -6,14 +6,18 @@ Change are listed in reverse chronological order (newest to oldest).
 
 <span class="changelog">
 
+###### [ 1.0.280 ] - 2026/07/26
+
+  * Updated `ZeroconfConnect` logic to add trace code for Spotify Connect Zeroconf `202 ERROR-LOGIN-FAILED` responses.
+
 ###### [ 1.0.279 ] - 2026/07/25
 
-  * Updated `ZeroconfConnect` logic to use a 10 second connection timeout value and a 30 second read timeout value for all `requests.post` calls. A local LAN connected device should establish a connection in well under 1 second. If a Spotify Connect device doesn't answer within 5–10 seconds, it's usually offline or unresponsive; once connected, the device should respond with data within 30 seconds - if not, then there is a problem.
+  * Updated `ZeroconfConnect` logic to use a 10 second connection timeout value and a 30 second read timeout value for all `requests.post` calls. A local LAN connected device should establish a connection in well under 1 second. If a Spotify Connect device doesn't answer within 5-10 seconds, it's usually offline or unresponsive; once connected, the device should respond with data within 30 seconds - if not, then there is a problem.
   * This release corrects v1.0.278, which contained an invalud UniCode character in some code comments.
 
 ###### [ 1.0.278 ] - 2026/07/25
 
-  * Updated `ZeroconfConnect` logic to use a 10 second connection timeout value and a 30 second read timeout value for all `requests.post` calls. A local LAN connected device should establish a connection in well under 1 second. If a Spotify Connect device doesn't answer within 5–10 seconds, it's usually offline or unresponsive; once connected, the device should respond with data within 30 seconds - if not, then there is a problem.
+  * Updated `ZeroconfConnect` logic to use a 10 second connection timeout value and a 30 second read timeout value for all `requests.post` calls. A local LAN connected device should establish a connection in well under 1 second. If a Spotify Connect device doesn't answer within 5-10 seconds, it's usually offline or unresponsive; once connected, the device should respond with data within 30 seconds - if not, then there is a problem.
 
 ###### [ 1.0.277 ] - 2026/07/21